requirements.txt
apache-airflow>=2.10.0
apache-airflow-providers-google>=8.1.0
apache-airflow-providers-apache-spark>=3.0.0
google-cloud-storage>=1.30.0
google-cloud-bigquery>=3.0.0
pyspark>=3.4.1
requests>=2.25.1
PyYAML>=6.0
pandas>=1.3.0
google-auth>=2.0.0
google-auth-oauthlib>=0.4.6
apache-airflow-providers-http>=3.0.0
python-dotenv>=0.21.0
